+++ cctools/file/softmagic.c
@@ -156,7 +156,7 @@ int nbytes;
 		if (magic[magindex].desc[0])
 			need_separator = 1;
 		/* and any continuations that match */
-		if (++cont_level >= tmplen)
+		if ((size_t)++cont_level >= tmplen)
 			if ((tmpoff = (int32 *) realloc(tmpoff,
 						       tmplen += 20)) == NULL)
 				error("out of memory\n");
@@ -200,7 +200,7 @@ int nbytes;
 					 * at a higher level,
 					 * process them.
 					 */
-					if (++cont_level >= tmplen)
+					if ((size_t)++cont_level >= tmplen)
 						if ((tmpoff = 
 						    (int32 *) realloc(tmpoff,
 						    tmplen += 20)) == NULL)
@@ -353,7 +353,7 @@ int nbytes;
 {
 	int32 offset = m->offset;
 
-	if (offset + sizeof(union VALUETYPE) <= nbytes)
+	if (offset + (int32)sizeof(union VALUETYPE) <= nbytes)
 		memcpy(p, s + offset, sizeof(union VALUETYPE));
 	else {
 		/*
@@ -389,7 +389,7 @@ int nbytes;
 			break;
 		}
 
-		if (offset + sizeof(union VALUETYPE) > nbytes)
+		if (offset + (int32)sizeof(union VALUETYPE) > nbytes)
 			return 0;
 
 		memcpy(p, s + offset, sizeof(union VALUETYPE));
